+++ pyodide_build/recipe.py
@@ -2,8 +2,8 @@ import functools
 from collections.abc import Iterable
 from pathlib import Path
 
-from .io import MetaConfig
-from .logger import logger
+from pyodide_build.io import MetaConfig
+from pyodide_build.logger import logger
 
 
 @functools.lru_cache(maxsize=1)
@@ -87,8 +87,9 @@ def load_recipes(
 
         elif name_or_tag in ("core", "min-scipy-stack"):
             logger.warning(
-                f"Using meta package without the 'tag:' prefix is deprecated,"
-                f" use 'tag:{name_or_tag}' instead."
+                "Using meta package without the 'tag:' prefix is deprecated, "
+                "use 'tag:%s' instead.",
+                name_or_tag,
             )
             for recipe in tagged_recipes[name_or_tag]:
                 recipes[recipe.package.name] = recipe.model_copy(deep=True)