+++ pyodide_build/xbuildenv.py
@@ -8,10 +8,10 @@ from urllib.request import urlopen
 
 from pyodide_lock import PyodideLockSpec
 
-from . import build_env
-from .create_package_index import create_package_index
-from .logger import logger
-from .xbuildenv_releases import (
+from pyodide_build import build_env
+from pyodide_build.create_package_index import create_package_index
+from pyodide_build.logger import logger
+from pyodide_build.xbuildenv_releases import (
     CrossBuildEnvReleaseSpec,
     cross_build_env_metadata_url,
     load_cross_build_env_metadata,
@@ -106,7 +106,7 @@ class CrossBuildEnvManager:
         version
             The version of xbuildenv to use.
         """
-        logger.info(f"Using Pyodide cross-build environment version: {version}")
+        logger.info("Using Pyodide cross-build environment version: %s", version)
 
         version_path = self._path_for_version(version)
         if not version_path.exists():
@@ -311,7 +311,8 @@ class CrossBuildEnvManager:
 
         if not lockfile_path.exists():
             logger.warning(
-                f"Pyodide lockfile not found at {lockfile_path}. Skipping PyPI index creation"
+                "Pyodide lockfile not found at %s. Skipping PyPI index creation",
+                lockfile_path,
             )
             return